# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mission Control API
  description: |
    Mission Control REST API for managing playbooks, notifications,
    configurations, views, and system resources.
  version: 1.0.0
  contact:
    name: Flanksource
    url: https://flanksource.com
servers:
  - url: /
    description: Current server
...
security:
  - BearerAuth: []
paths:
...
  /rbac/{id}/update_role:
    post:
      tags: [RBAC]
      summary: Update roles for user
      description: Update the roles assigned to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          $ref: "#/components/responses/HTTPSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /rbac/{id}/roles:
    get:
      tags: [RBAC]
      summary: Get roles for user
      description: Get roles assigned to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User roles
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/HTTPSuccess"
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          type: string
        "500":
          $ref: "#/components/responses/InternalError"
  /rbac/token/{id}/permissions:
    get:
      tags: [RBAC]
      summary: Get permissions for token
      description: Get permissions assigned to an access token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Token permissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/HTTPSuccess"
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          $ref: "#/components/schemas/Permission"
        "500":
          $ref: "#/components/responses/InternalError"

  # Agent endpoints
...
  /playbook/list:
    get:
      tags: [Playbooks]
      summary: List playbooks
      description: List playbooks available for a config, check, or component
      parameters:
        - name: config_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by config ID
        - name: check_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by check ID
        - name: component_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by component ID
      responses:
        "200":
          description: List of playbooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybookListItem"
        "400":
          $ref: "#/components/responses/BadRequest"
  /playbook/webhook/{webhook_path}:
    post:
      tags: [Playbooks]
      summary: Trigger playbook via webhook
      description: Trigger a playbook run via webhook
      security:
        - {}
        - BasicAuth: []
        - BearerAuth: []
      parameters:
        - name: webhook_path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
        "401":
          $ref: "#/components/responses/Unauthorized"
  /playbook/{id}/params:
    post:
      tags: [Playbooks]
      summary: Get playbook parameters
      description: Get templated parameters for a playbook run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunParams"
      responses:
        "200":
          description: Playbook parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetParamsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /playbook/run:
    post:
      tags: [Playbooks]
      summary: Run playbook
      description: Start a new playbook run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunParams"
      responses:
        "201":
          description: Playbook run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (for webhooks)
  responses:
    HTTPSuccess:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPSuccess"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
  schemas:
    HTTPSuccess:
      type: object
      properties:
        message:
          type: string
        payload: {}
    HTTPError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        data: {}
...
    Permission:
      type: object
      properties:
        subject:
          type: string
        object:
          type: string
        action:
          type: string
        deny:
          type: boolean
...
    PlaybookListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        title:
          type: string
        icon:
          type: string
        description:
          type: string
        category:
          type: string
    RunParams:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        config_id:
          type: string
          format: uuid
        check_id:
          type: string
          format: uuid
        component_id:
          type: string
          format: uuid
        params:
          type: object
          additionalProperties:
            type: string
    RunResponse:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
    GetParamsResponse:
      type: object
      properties:
        params:
          type: array
          items:
            $ref: "#/components/schemas/PlaybookParameter"
    PlaybookParameter:
      type: object
      properties:
        name:
          type: string
        default:
          type: string
        label:
          type: string
        required:
          type: boolean
        icon:
          type: string
        description:
          type: string
        type:
          type: string
...